requirements.txt
streamlit==1.35.0
python-docx
PyPDF2
openai
python-dotenv